Rename control.h to mccontrol.h.
This is to avoid a naming conflict with control.h from the DirectShow
API.
