+++ engine/src/player-legacy.h
@@ -20,7 +20,7 @@ along with LiveCode.  If not see <http://www.gnu.org/licenses/>.  */
 #ifndef	PLAYER_LEGACY_H
 #define	PLAYER_LEGACY_H
 
-#include "control.h"
+#include "mccontrol.h"
 
 #ifdef FEATURE_QUICKTIME
 enum QTstate {
