+++ engine/src/redraw.cpp
@@ -27,7 +27,7 @@ along with LiveCode.  If not see <http://www.gnu.org/licenses/>.  */
 #include "stack.h"
 #include "stacklst.h"
 #include "card.h"
-#include "control.h"
+#include "mccontrol.h"
 #include "objptr.h"
 #include "group.h"
 #include "field.h"
