+++ engine/src/scrolbar.h
@@ -20,7 +20,7 @@ along with LiveCode.  If not see <http://www.gnu.org/licenses/>.  */
 #ifndef	SCROLLBAR_H
 #define	SCROLLBAR_H
 
-#include "control.h"
+#include "mccontrol.h"
 
 #define MIN_THUMB_SIZE 8
